UPSTREAM: <drop>: Use Clone for CloneTLSConfig for golang 1.8
Use Clone() for the implementation of CloneTLSConfig when golang version
is 1.8, otherwise continue to use existing implementation. This avoids
issues related Clone() calling serverInit on the source object if it was
not already previously run.

This ensures that CloneTLSConfig is able to be used consistently for
both golang 1.7 and golang 1.8.

This commit can be dropped when rebased on k8s 1.7, which will require
golang 1.8. Alternatively, this commit can be dropped if we move to
golang 1.8 exclusively and the related upstream Clone() changes are
backported.

:000000 100644 0000000000... f1f6193f9c... A	staging/src/k8s.io/apimachinery/pkg/util/net/clone.go
:000000 100644 0000000000... 49a2597b94... A	staging/src/k8s.io/apimachinery/pkg/util/net/clone_17.go
:000000 100644 0000000000... 77d40f4ac7... A	staging/src/k8s.io/apimachinery/pkg/util/net/clone_17_test.go
:000000 100644 0000000000... 66eafcaede... A	staging/src/k8s.io/apimachinery/pkg/util/net/clone_test.go
:100644 100644 3842ca0eeb... c081bd0892... M	staging/src/k8s.io/apimachinery/pkg/util/net/http.go
:100644 100644 30cbb6edc4... 8b0e858c1d... M	staging/src/k8s.io/apimachinery/pkg/util/net/http_test.go
:100644 100644 ee143b1e24... 9b2e0c93ed... M	staging/src/k8s.io/apiserver/pkg/util/proxy/dial_test.go
