+++ staging/src/k8s.io/apimachinery/pkg/util/net/http_test.go
@@ -1,5 +1,3 @@
-// +build go1.8
-
 /*
 Copyright 2016 The Kubernetes Authors.
 
@@ -25,71 +23,8 @@ import (
 	"net/url"
 	"os"
 	"reflect"
-	"runtime"
-	"strings"
 	"testing"
-
-	"k8s.io/apimachinery/pkg/util/sets"
 )
-
-func TestCloneTLSConfig(t *testing.T) {
-	expected := sets.NewString(
-		// These fields are copied in CloneTLSConfig
-		"Rand",
-		"Time",
-		"Certificates",
-		"RootCAs",
-		"NextProtos",
-		"ServerName",
-		"InsecureSkipVerify",
-		"CipherSuites",
-		"PreferServerCipherSuites",
-		"MinVersion",
-		"MaxVersion",
-		"CurvePreferences",
-		"NameToCertificate",
-		"GetCertificate",
-		"ClientAuth",
-		"ClientCAs",
-		"ClientSessionCache",
-
-		// These fields are not copied
-		"SessionTicketsDisabled",
-		"SessionTicketKey",
-
-		// These fields are unexported
-		"serverInitOnce",
-		"mutex",
-		"sessionTicketKeys",
-
-		// go1.8
-		"DynamicRecordSizingDisabled",
-		"GetClientCertificate",
-		"GetConfigForClient",
-		"KeyLogWriter",
-		"Renegotiation",
-		"VerifyPeerCertificate",
-		"originalConfig",
-	)
-
-	// See #33936.
-	if strings.HasPrefix(runtime.Version(), "go1.7") {
-		expected.Insert("DynamicRecordSizingDisabled", "Renegotiation")
-	}
-
-	fields := sets.NewString()
-	structType := reflect.TypeOf(tls.Config{})
-	for i := 0; i < structType.NumField(); i++ {
-		fields.Insert(structType.Field(i).Name)
-	}
-
-	if missing := expected.Difference(fields); len(missing) > 0 {
-		t.Errorf("Expected fields that were not seen in http.Transport: %v", missing.List())
-	}
-	if extra := fields.Difference(expected); len(extra) > 0 {
-		t.Errorf("New fields seen in http.Transport: %v\nAdd to CopyClientTLSConfig if client-relevant, then add to expected list in TestCopyClientTLSConfig", extra.List())
-	}
-}
 
 func TestGetClientIP(t *testing.T) {
 	ipString := "10.0.0.1"
